requirements.txt
flask
flask_sqlalchemy
werkzeug
python-dotenv
google-generativeai
ffmpeg-python
